requirements.txt
langchain
langchain-openai
google-api-python-client
google-auth
google-auth-oauthlib

chardet
markdown

flask
flask-cors
python-dotenv
gunicorn


